api_server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.91",
    "fastapi[standard]>=0.135.3",
    "opentelemetry-api>=1.41.1",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-instrumentation-fastapi>=0.62b1",
    "opentelemetry-instrumentation-redis>=0.62b1",
    "opentelemetry-instrumentation-botocore>=0.62b1",
    "opentelemetry-sdk>=1.41.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "python-logging-loki-v2>=1.1.1",
    "redis>=7.4.0",
    "ruff>=0.15.11",
    "valkey[libvalkey]>=6.1.1",
]
